fix: inclusion proof hash shall have reverse order
