+++ src/app/components/Connect/ConnectedSmall.tsx
@@ -5,7 +5,7 @@ import { useOnClickOutside } from "usehooks-ts";
 
 import { useAppState } from "@/app/state";
 import { getNetworkConfig } from "@/config/network.config";
-import { satoshiToBtc } from "@/utils/btcConversions";
+import { satoshiToBtc } from "@/utils/btc";
 import { maxDecimals } from "@/utils/maxDecimals";
 import { trim } from "@/utils/trim";
