+++ src/app/components/Stakers/Staker.tsx
@@ -1,5 +1,5 @@
 import { getNetworkConfig } from "@/config/network.config";
-import { satoshiToBtc } from "@/utils/btcConversions";
+import { satoshiToBtc } from "@/utils/btc";
 import { maxDecimals } from "@/utils/maxDecimals";
 
 import { Hash } from "../Hash/Hash";
