+++ src/app/components/Staking/Form/StakingFee.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 
 import { getNetworkConfig } from "@/config/network.config";
-import { satoshiToBtc } from "@/utils/btcConversions";
+import { satoshiToBtc } from "@/utils/btc";
 import { getFeeRateFromMempool } from "@/utils/getFeeRateFromMempool";
 import { Fees } from "@/utils/wallet/btc_wallet_provider";
