+++ src/utils/btc.ts
@@ -17,3 +17,16 @@ export function satoshiToBtc(satoshi: number): number {
 export function btcToSatoshi(btc: number): number {
   return Math.round(btc * 1e8);
 }
+
+/**
+ * Derives the merkle proof from the list of hex strings. Note the
+ * sibling hashes are reversed from hex before concatenation.
+ * @param merkle - The merkle proof hex strings.
+ * @returns The merkle proof in hex string format.
+ */
+export const deriveMerkleProof = (merkle: string[]) => {
+  const proofHex = merkle.reduce((acc: string, m: string) => {
+    return acc + Buffer.from(m, "hex").reverse().toString("hex");
+  }, "");
+  return proofHex;
+};
