+++ tests/utils/btcConvensions.test.ts
@@ -1,4 +1,4 @@
-import { btcToSatoshi, satoshiToBtc } from "@/utils/btcConversions";
+import { btcToSatoshi, satoshiToBtc } from "@/utils/btc";
 
 describe("satoshiToBtc", () => {
   it("should convert 0 satoshis to 0 BTC", () => {
